+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Configuration option for artist separator characters `gui.library.artist_separator`
 - Docs subpage for configuration (including content)
+- `typing_extensions` is now a dependency, to allow for more typing features
+- The model api routes now allows for `DELETE` requests to delete resources by id. Not used yet but will be helpful for future features.
 
 ### Fixed
 
@@ -21,6 +23,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cache invalidation now triggers on delete folder in frontend [#138](https://github.com/pSpitzner/beets-flask/issues/138)
 - In albums and items view the clicking on artists does not return any results if the contained a separator character (e.g. `&`) [#132](https://github.com/pSpitzner/beets-flask/issues/138)
 - Cleanup old actions.tsx file, which included old unused code [#134](https://github.com/pSpitzner/beets-flask/issues/134)
+
+### Changed
+
+- Created `types.py` file to hold custom sqlalchemy types, and moved `IntDictType` there.
 
 ## [1.0.0] - 25-07-06